% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbscan_SH.R
\name{dbscan_SH}
\alias{dbscan_SH}
\title{dbscan_SH}
\usage{
dbscan_SH(data, eps = NULL, showplot = F, prop_outliers = 0.1,
  eps_res = 500, eps_range = NULL)
}
\arguments{
\item{data}{$Y of Rtsne output}

\item{showplot}{Logical. TRUE to plot silhouette score and eps. Default is FALSE.}

\item{prop_outliers}{A value between 0 to 1. The proportion of data points allowed to be unclustered. Default is 0.1.}

\item{eps_res}{Integer. The resolution of simulation. Default is 500.}

\item{eps_range}{NULL or vectore of two values. The space of search for eps parameter. If a vectore is given, simulation is performed between the given range. Otherwise, te serch space is defined based on input data size.}
}
\value{
A vector of cluster label. 0 is for outlyers.
}
\description{
DBSCAN with silhouette score optimization
}

