% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meredith.R
\name{meredith}
\alias{meredith}
\title{MEREDITH (multi omic data integration approach)}
\usage{
meredith(data, scale = FALSE, transfer = FALSE, takepc = 50,
  explainedVar = FALSE, nTSNE = 100, dims = 2, initial_dims = 50,
  perplexity = 30, theta = 0.5, max_iter = 1000)
}
\arguments{
\item{data}{List of matrix or data frame. (Note that even when input is one matrix/dataframe, use list())}

\item{transfer}{Logical. TRUE if the input matrix/data frame have to be transfered. Default is FALSE.}

\item{takepc}{Integer. The number of PC to keep. Default is 50.}

\item{nTSNE}{Integer. The number of run for tSNE. Default is 100.}

\item{dims, initial_dims, preplexity, theta, max_iter}{option of Rtsne (see ?Rtsne).}

\item{Logical.}{The scale option for PCA (see ?prcomp). Default is FALSE}

\item{explainedvar}{FALSE or a value. The percentage of explained variance to keep. Default is FALSE. If this argument is value between 0 and 100, takepc will be ignored.}
}
\value{
list of data same as the Rtsne output (see ?Rtsne for details)
}
\description{
MEREDITH combines a list of input data (data.frame or matrix) and perform tSNE
}

