% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{meredith}
\alias{meredith}
\title{MEREDITH function}
\usage{
meredith(data, scale = FALSE, transfer = FALSE, takepc = 50,
  explainedVar = FALSE, nTSNE = 100, dims = 2, initial_dims = 50,
  perplexity = 30, theta = 0.5, max_iter = 1000)
}
\arguments{
\item{data}{list of matrix or data frame (Note that even when input is one matrix/dataframe, use list())}

\item{scale}{option for PCA (see ?prcomp). Default is FALSE}

\item{transfer}{logical. TRUE if the input matrix/data frame have to be transfered. Default is FALSE.}

\item{takepc}{the number of PC to keep. Default is 50.}

\item{nTSNE}{The number of run for tSNE. Default is 100.}

\item{dims, initial_dims, preplexity, theta, max_iter}{option of Rtsne (see ?Rtse).}

\item{explainedvar}{FALSE or percentage of explained variance to keep. Default is FALSE. If this argument is value between 0 and 100, takepc will be ignored.}
}
\value{
list same as Rtsne output (see ?Rtsne for details)
export
}
\description{
MEREDITH function
}

